refactor(has_method_result_steps): extract services to avoid circular dependency + use std config
