+++ lib/convenient_service/services.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
-require_relative "services/method_step_config"
 require_relative "services/run_own_method_in_organizer"
 require_relative "services/run_method_in_organizer"
